added middlewares for role permissions

+++ routes/web.php
@@ -22,19 +22,15 @@ Route::get('/', function () {
     return view('landing');
 });
 
-Route::get('/users/register', [UserController::class, 'register']);
+Route::get('/users/register', [UserController::class, 'register'])->middleware('guest');
 
 Route::post('/register', [UserController::class, 'store'])->middleware('guest');
 
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 Route::get('/users/sign_in', [UserController::class, 'sign_in'])->name('login')->middleware('guest');
 
-Route::post('/sign_in', [UserController::class, 'authenticate']);
-
-Route::get('/sign_in', function () {
-    return view('sign_in');
-});
+Route::post('/sign_in', [UserController::class, 'authenticate'])->middleware('guest');
 
 Route::get('/dashboard', function () {
     return view('dashboard', [
@@ -45,7 +41,7 @@ Route::get('/dashboard', function () {
         ],
         'title' => 'Dashboard | ApexHubSpot'
     ]);
-})/*->middleware('auth')*/;
+})->middleware('auth');
 
 Route::get('/office_map', function () {
     return view('office_map', [
@@ -54,7 +50,7 @@ Route::get('/office_map', function () {
         ],
         'title' => 'Office Map | ApexHubSpot'
     ]);
-})/*->middleware('auth')*/;
+})->middleware('auth');
 
 Route::get('/faqs', function () {
     return view('faqs', [
@@ -65,7 +61,7 @@ Route::get('/faqs', function () {
         ],
         'title' => 'Frequently Asked Questions | ApexHubSpot'
     ]);
-})/*->middleware('auth')*/;
+})->middleware('auth');
 
 Route::get('/guide', function () {
     return view('guide', [
@@ -77,7 +73,7 @@ Route::get('/guide', function () {
         ],
         'title' => 'User Guide | ApexHubSpot'
     ]);
-})/*->middleware('auth')*/;
+})->middleware('auth');
 
 Route::get('/profile', function () {
     return view('profile', [
@@ -89,7 +85,7 @@ Route::get('/profile', function () {
         ],
         'title' => 'Profile | ApexHubSpot'
     ]);
-})/*->middleware('auth')*/;
+})->middleware('auth');
 
 // Common Resource Routes:
 // index - Show all items
@@ -100,33 +96,33 @@ Route::get('/profile', function () {
 // update - Update item
 // destroy - Delete item
 
-Route::get('/users', [UserController::class, 'index'])/*->middleware('auth')*/;
+Route::get('/users', [UserController::class, 'index'])->middleware(['auth', 'admin']);
 
 // for creating item
-Route::get('/users/create', [UserController::class, 'create'])/*->middleware('auth')*/;
+Route::get('/users/create', [UserController::class, 'create'])->middleware(['auth', 'admin']);
 
-Route::post('/users', [UserController::class, 'admin_store'])/*->middleware('auth')*/;
+Route::post('/users', [UserController::class, 'admin_store'])->middleware(['auth', 'admin']);
 
 // for editing item
-Route::get('/users/{user}/edit', [UserController::class, 'edit'])/*->middleware('auth')*/;
+Route::get('/users/{user}/edit', [UserController::class, 'edit'])->middleware(['auth', 'admin']);
 
-Route::put('/users/{user}/edit', [UserController::class, 'update'])/*->middleware('auth')*/;
+Route::put('/users/{user}/edit', [UserController::class, 'update'])->middleware(['auth', 'admin']);
 
-Route::put('/users/{user}', [UserController::class, 'approve'])/*->middleware('auth')*/;
+Route::put('/users/{user}', [UserController::class, 'approve'])->middleware(['auth', 'office_manager']);
 
-Route::delete('/users/{user}', [UserController::class, 'destroy'])/*->middleware('auth')*/;
+Route::delete('/users/{user}', [UserController::class, 'destroy'])->middleware(['auth', 'admin']);
 
-Route::get('/desks', [DeskController::class, 'index'])/*->middleware('auth')*/;
+Route::get('/desks', [DeskController::class, 'index'])->middleware(['auth', 'office_manager']);
 
-Route::get('/desks/create', [DeskController::class, 'create'])/*->middleware('auth')*/;
+Route::get('/desks/create', [DeskController::class, 'create'])->middleware(['auth', 'office_manager']);
 
-Route::post('/desks', [DeskController::class, 'store'])/*->middleware('auth')*/;
+Route::post('/desks', [DeskController::class, 'store'])->middleware(['auth', 'office_manager']);
 
-Route::put('/desks/{desk}', [DeskController::class, 'availability'])/*->middleware('auth')*/;
+Route::put('/desks/{desk}', [DeskController::class, 'availability'])->middleware(['auth', 'office_manager']);
 
-Route::delete('/desks/{desk}', [DeskController::class, 'destroy'])/*->middleware('auth')*/;
+Route::delete('/desks/{desk}', [DeskController::class, 'destroy'])->middleware(['auth', 'office_manager']);
 
-Route::get('/bookings', [BookingController::class, 'index'])/*->middleware('auth')*/;
+Route::get('/bookings', [BookingController::class, 'index'])->middleware('auth');
 
 // * UNUSED ROUTES
 
@@ -139,9 +135,9 @@ Route::get('/bookings', [BookingController::class, 'index'])/*->middleware('auth
 //         ],
 //         'title' => 'Manage Roles | ApexHubSpot'
 //     ]);
-// })/*->middleware('auth')*/;
+// })->middleware('auth');
 
-// Route::get('/bookings/history', [BookingController::class, 'history'])/*->middleware('auth')*/;
+// Route::get('/bookings/history', [BookingController::class, 'history'])->middleware('auth');
 
 // Show route should be always at the last line after preceeding paths
 // Route::get('/users/{user}', [UserController::class, 'show'])->middleware('auth');
